fix: updated dependencies to match version used in apps; refactor: moved ShortcutLinks component to a dedicated folder;
