+++ packages/extension/src/newtab/ShortcutLinks/ShortcutLinks.spec.tsx
