+++ packages/storybook/package.json
@@ -35,9 +35,9 @@
     "@storybook/react": "^8.1.11",
     "@storybook/react-vite": "^8.1.11",
     "@storybook/theming": "^8.1.11",
-    "@svgr/plugin-jsx": "^8.1.0",
-    "@svgr/plugin-svgo": "^8.1.0",
-    "@types/node": "^16.11.41",
+    "@svgr/plugin-jsx": "^5.5.0",
+    "@svgr/plugin-svgo": "^5.5.0",
+    "@types/node": "^20.14.9",
     "@types/react": "^18.3.1",
     "@types/webextension-polyfill": "^0.10.5",
     "@vitejs/plugin-react": "^1.3.2",
@@ -56,7 +56,7 @@
     "postcss-simple-vars": "^6.0.3",
     "serve": "^13.0.4",
     "storybook": "^8.1.11",
-    "typescript": "5.3.3",
+    "typescript": "^4.9.5",
     "vite": "^5.3.3",
     "vite-plugin-svgr": "^4.2.0"
   }
